+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `participant-activity` and `participant-module` middleware groups
 - Migration to update settings
 - Added a title prefix to resource
+- Name the main module route 'module'
 
 ## [5.0.0] - (11/08/2021)